pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "shesha"
dynamic = ["version"]
description = "Recursive Language Models for document querying"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Shesha Authors" }]

dependencies = [
    "litellm>=1.0",
    "docker>=7.0",
    "pyyaml>=6.0",
    "pdfplumber>=0.10",
    "python-docx>=1.0",
    "beautifulsoup4>=4.12",
    "chardet>=5.0",
    "openpyxl>=3.1",
    "python-pptx>=1.0",
    "striprtf>=0.0.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=6.0",
    "ruff>=0.1",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
    "types-docker>=7.0",
    "textual>=1.0",
    "datasets",
    "pandas",
    "tqdm",
    "matplotlib",
    "python-dateutil",
    "arxiv>=2.0",
    "bibtexparser>=2.0.0b7",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
    "httpx>=0.27",
]
arxiv = [
    "arxiv>=2.0",
    "bibtexparser>=2.0.0b7",
    "httpx>=0.27",
]
tui = [
    "textual>=1.0",
]
web = [
    "shesha[arxiv]",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
    "httpx>=0.27",
]

[project.scripts]
shesha-web = "shesha.experimental.web.__main__:main"
shesha-code = "shesha.experimental.code_explorer.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/shesha"]

[tool.hatch.build.targets.wheel.force-include]
"prompts" = "src/shesha/prompts/prompts"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/shesha/_version.py"

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["src/shesha/_version.py"]  # Auto-generated by setuptools-scm

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/shesha/rlm/prompts.py" = ["E501"]  # Long lines in prompt templates are fine

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "arxiv"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "uvicorn"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "openpyxl"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "striprtf.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning:litellm.*:",  # litellm uses deprecated asyncio.iscoroutinefunction
]
